add select map option

+++ menu/editor_menu/ui/ui.gd
@@ -4,8 +4,10 @@ onready var _maps :Array = Utils.load_maps() # [ HexMapFileManifest ]
 
 onready var map_list = $CanvasLayer/Control/SafeArea/VBoxContainer/HBoxContainer/ScrollContainer/map_list
 onready var new_map_dialog = $CanvasLayer/Control/new_map_dialog
+onready var dialog_confirm = $CanvasLayer/Control/dialog_confirm
 
 func _ready():
+	dialog_confirm.visible = false
 	new_map_dialog.visible = false
 	show_maps()
 	
@@ -27,12 +29,16 @@ func _on_edit_map(data :HexMapFileManifest):
 	load_map(data.map_file_path)
 
 func _on_delete_map(data :HexMapFileManifest):
-	var path = "user://%s/%s.manifest" % [Global.map_dir, data.map_name]
-	SaveLoad.delete_save(data.map_file_path, false)
-	SaveLoad.delete_save(data.map_image, false)
-	SaveLoad.delete_save(path, false)
-	_maps = Utils.load_maps()
-	show_maps()
+	dialog_confirm.set_message("Are you sure want delete %s ?" % data.map_name)
+	dialog_confirm.visible = true
+	var ok = yield(dialog_confirm,"close")
+	if ok:
+		var path = "user://%s/%s.manifest" % [Global.map_dir, data.map_name]
+		SaveLoad.delete_save(data.map_file_path, false)
+		SaveLoad.delete_save(data.map_image, false)
+		SaveLoad.delete_save(path, false)
+		_maps = Utils.load_maps()
+		show_maps()
 	
 func load_map(filename :String):
 	var can_load = Global.load_map(filename, false)
@@ -45,7 +51,6 @@ func load_map(filename :String):
 func _on_new_map_pressed():
 	new_map_dialog.visible = true
 	
-
 func _on_search_map_text_changed(new_text):
 	show_maps(new_text)
